% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_ingredients_graph.R
\name{clean_ingredients}
\alias{clean_ingredients}
\title{Cleans ingredient data, combines the same ingredients that have slightly different names
This fixes the count for graph, all instances are counted together instead of separately}
\usage{
clean_ingredients(x)
}
\arguments{
\item{x}{Vector of ingredient names}
}
\description{
Cleans ingredient data, combines the same ingredients that have slightly different names
This fixes the count for graph, all instances are counted together instead of separately
}
