% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_recipes_graph.R
\name{compare_recipes_graph}
\alias{compare_recipes_graph}
\title{Provide the recipe numbers for the two recipes being compared}
\usage{
compare_recipes_graph(x, recipe1, recipe2)
}
\arguments{
\item{x}{Dataset of recipes}

\item{recipe1}{Row number of the first recipe}

\item{recipe2}{Row number of the second recipe}
}
\description{
Provide the recipe numbers for the two recipes being compared
}
