% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_recipe.R
\name{find_recipe}
\alias{find_recipe}
\title{Searches for a specific recipe based on the users input, whether it be the name of a meal or a singular food item.
Outputs the name of the recipes and their number.}
\usage{
find_recipe(x, input)
}
\arguments{
\item{x}{Dataset of recipes}

\item{input}{Keyword or phrase to search for in recipe titles}
}
\description{
Searches for a specific recipe based on the users input, whether it be the name of a meal or a singular food item.
Outputs the name of the recipes and their number.
}
