% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe.R
\name{get_recipe}
\alias{get_recipe}
\title{After using find_recipe, which gives the name of recipes and the recipe number, use this function to get
the ingredients, directions, and link for one recipe. Must use the recipe number because some recipes have the same name.}
\usage{
get_recipe(x, index)
}
\arguments{
\item{x}{Dataset of recipes}

\item{index}{Row number of the recipe}
}
\description{
After using find_recipe, which gives the name of recipes and the recipe number, use this function to get
the ingredients, directions, and link for one recipe. Must use the recipe number because some recipes have the same name.
}
