% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_recipe.R
\name{random_recipe}
\alias{random_recipe}
\title{Finds a random recipe and returns the entire row of information. Title, ingredients, directions, and link.}
\usage{
random_recipe(x)
}
\arguments{
\item{x}{Dataset of recipes}
}
\description{
Finds a random recipe and returns the entire row of information. Title, ingredients, directions, and link.
}
