% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_ingredients_graph.R
\name{top_ingredients_graph}
\alias{top_ingredients_graph}
\title{Creates a horizontal bar graph of the top 10 or (user chooses), ingredients used in recipes
Counts how many recipes each ingredient is in, then shows it in decreasing order}
\usage{
top_ingredients_graph(x, top_n = 10)
}
\arguments{
\item{x}{Dataset of recipes}

\item{top_n}{Number of top ingredients to display}
}
\description{
Creates a horizontal bar graph of the top 10 or (user chooses), ingredients used in recipes
Counts how many recipes each ingredient is in, then shows it in decreasing order
}
