% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_to_make_with_ingredients.R
\name{what_to_make_with_ingredients}
\alias{what_to_make_with_ingredients}
\title{This function allows the user to filter recipes by their ingredients. Finds recipes that include all of the ingredients you want to use.}
\usage{
what_to_make_with_ingredients(x, input)
}
\arguments{
\item{x}{Dataset of recipes}

\item{input}{Vector of ingredient names to search for}
}
\description{
This function allows the user to filter recipes by their ingredients. Finds recipes that include all of the ingredients you want to use.
}
